move reminder to serverext
